Revert "Corrección atributo"
This reverts commit 71c84d57d3a83b7434bdb2e3a7d58704d799e374.

+++ src/main/java/com/mycompany/ricettapp/archivos/GestorJSONv2.java
@@ -141,8 +141,8 @@ public class GestorJSONv2 {
             String votos = ((String) jsonObject.get("votos"));
 
             r.setNombre(nombre);
-            r.setRanking(Integer.parseInt(ranking));
-            r.setVotos(Integer.parseInt(votos));
+            r.setRanking(Double.parseDouble(ranking));
+            r.setVotos(Double.parseDouble(votos));
 
             ingredientes(jsonObject, ingredientesAL, r);
             instrucciones(jsonObject, instruccionesAL, r);
